Updates to consensus_enhancer regions and writeup for consensus_open_chromatin_regions
Updated consensus_enhancer_regions dataset to include hg38 liftover for the eRNA files. In addition I updated both this script and the script for consensus_open_chromatin_regions to allow the download and import of the chain file on non-unix based systems (the consensus_open_chromatin_dataset however remains unchanged).